Create README.md
Howdy, I am focused to get cliams and applications processed with various agencies to include: Start Retiretment, SSA, JFS, IRS, LIFELINE, OHIO BENEFITS, CREDIT REPORT AND SECURITY , HEALTHCARE PLANS   --claim appeals, Disability, Health Plans, 401k investing, benefits that are needed and eligible elders and others who have no computer skills and stop the applying process because its to difficult. When they are entitled and eligible and its there for the reason to claim and I am good with  staying on top and appeals or writing narratives explanaations or what would be the program that is needed for that specific person, I am one that likes to find every way to get what a humn deserves to feel comfortable and im annoying to the point of getting it done fast bc i will ty on top of it. its deserved and these agenies are here for us and when needed, I will help if needed. gladly. i fight because one small things leads to more and can be a start to get people working or in any stage of life get prepared for next step. I enjoy helping people instead of them not trying or giving up as my clients say because confusing or to difficult. no more struggling. Use me as a stepping stone. no fee or charge.